change hash from size_t to uint32_t

+++ concurrent/project/inc/hashtable.h
@@ -42,7 +42,7 @@ typedef void * hashtable_elem_t;
 /**
  * @brief   Function signature for hashing key objects
  */
-typedef size_t (*hash_f_t)(hashtable_key_t);
+typedef uint32_t (*hash_f_t)(hashtable_key_t);
 
 /* --- PUBLIC FUNCTION PROTOTYPES ------------------------------------------- */
